+++ appland-user-milestones/app.js
@@ -1,0 +1,6 @@
+import Vue from 'vue';
+import { default as plugin } from '@appland/components'; // eslint-disable-line import/no-named-default
+
+export { mountQuickstartInstallAgent, mountQuickstartOpenAppmaps } from './quickstartDocsView';
+
+Vue.use(plugin);