+++ src/main/java/appland/actions/OpenQuickstartAction.java
@@ -1,0 +1,18 @@
+package appland.actions;
+
+import appland.milestones.UserMilestonesEditorProvider;
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.diagnostic.Logger;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Objects;
+
+public class OpenQuickstartAction extends AnAction {
+    private static final Logger LOG = Logger.getInstance("#appmap.action");
+
+    @Override
+    public void actionPerformed(@NotNull AnActionEvent e) {
+        UserMilestonesEditorProvider.openUserMilestones(Objects.requireNonNull(e.getProject()));
+    }
+}