+++ src/main/java/appland/settings/AppMapApplicationSettingsService.java
@@ -5,7 +5,6 @@ import com.intellij.openapi.components.PersistentStateComponent;
 import com.intellij.openapi.components.State;
 import com.intellij.openapi.components.Storage;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 @State(name = "appmap-application", storages = @Storage("appmap-application.xml"))
 public class AppMapApplicationSettingsService implements PersistentStateComponent<AppMapApplicationSettings> {
@@ -13,16 +12,16 @@ public class AppMapApplicationSettingsService implements PersistentStateComponen
 
     @NotNull
     public static AppMapApplicationSettings getInstance() {
-        return ApplicationManager.getApplication().getService(AppMapApplicationSettingsService.class).state;
+        return ApplicationManager.getApplication().getService(AppMapApplicationSettingsService.class).getState();
     }
 
     @Override
-    public @Nullable AppMapApplicationSettings getState() {
+    public synchronized @NotNull AppMapApplicationSettings getState() {
         return state;
     }
 
     @Override
-    public void loadState(@NotNull AppMapApplicationSettings state) {
+    public synchronized void loadState(@NotNull AppMapApplicationSettings state) {
         this.state = state;
     }
 }